requirements.txt
# Core Dependencies
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
ultralytics>=8.0.0

# Image Processing & Data Augmentation
opencv-python==4.8.1.78
Pillow==10.1.0
numpy>=1.24.0
albumentations>=1.3.0

# Data Science & ML
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
tqdm>=4.65.0

# System & Utils
psutil>=5.9.0

# Configuration
PyYAML==6.0.1

# Testing
pytest==7.4.3
pytest-qt==4.2.0
pytest-cov==4.1.0
pytest-benchmark>=4.0.0

# Development Tools (Optional)
black==23.12.0
flake8==6.1.0
mypy==1.7.1
